feat: Add word detail ui (#14)
word detail ui 구성, 파일 정리
