+++ Android/app/src/main/AndroidManifest.xml
@@ -52,19 +52,19 @@
             android:name=".presentation.my.MyWordActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.learn.LearnSyllableDetailActivity"
+            android:name=".presentation.learn.syllable.LearnSyllableDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.learn.LearnPhonemeDetailActivity"
+            android:name=".presentation.learn.phoneme.LearnPhonemeDetailActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.learn.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.learn.LearnSyllableActivity"
+            android:name=".presentation.learn.syllable.LearnSyllableActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.learn.LearnPhonemeActivity"
+            android:name=".presentation.learn.phoneme.LearnPhonemeActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.onboarding.OnBoardingFinishActivity"
