+++ Android/app/src/main/res/layout/activity_learn_phoneme_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.learn.LearnPhonemeDetailActivity">
+    tools:context=".presentation.learn.phoneme.LearnPhonemeDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_phoneme"
