+++ Android/app/src/main/res/layout/activity_learn_syllable_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.learn.LearnSyllableDetailActivity">
+    tools:context=".presentation.learn.syllable.LearnSyllableDetailActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_phoneme"
