+++ Android/app/src/main/res/layout/activity_learn_word_detail.xml
@@ -28,12 +28,39 @@
         android:layout_marginStart="30dp"/>
 
     <TextView
+        android:id="@+id/ic_word_detail"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="🔗"
+        android:textSize="25sp"
+        app:layout_constraintTop_toTopOf="@+id/tv_word_detail_main"
+        app:layout_constraintBottom_toBottomOf="@+id/tv_word_detail_main"
+        app:layout_constraintEnd_toStartOf="@+id/tv_word_detail_main"
+        android:layout_marginEnd="5dp"/>
+
+    <TextView
         android:id="@+id/tv_word_detail_main"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text=""
+        android:text="ㄱ"
+        android:textStyle="bold"
+        android:textSize="25sp"
+        android:textColor="@color/font"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tv_word_detail_title"
-        android:layout_marginTop="20dp" />
+        android:layout_marginTop="30dp" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rv_word_detail"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@+id/tv_word_detail_main"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginTop="20dp"
+        app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+        app:spanCount="3"
+        tools:listitem="@layout/word_detail_item" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
