+++ Android/app/src/main/res/layout/fragment_learn_phoneme_consonant.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.learn.LearnPhonemeConsonantFragment"
+    tools:context=".presentation.learn.phoneme.LearnPhonemeConsonantFragment"
     android:layout_gravity="center">
 
     <!-- TODO: Update blank fragment layout -->
