+++ Android/app/src/main/res/layout/word_detail_item.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:padding="8dp">
+
+    <Button
+        android:id="@+id/btn_syllable"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="가위"
+        android:background="@drawable/btn_background_round_border"
+        android:textColor="@color/main"
+        android:textSize="18sp"
+        android:paddingVertical="15dp"
+        android:paddingHorizontal="35dp"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
